Fixed bugs in bash_profile and bashrc--
When tmux started for the first time, it would display errant processes;
restricted the PPID to 1 filtered out those; not sure if PPID will always
be 1, so will have to wait and see.  Also, seems to be a kind of bug in
Bash with the builtin test command, using -n on an empty or unset variable;
it always produced true, when it should have produced false; -z works
correctly, and [[ -n  ]] works correctly; switched to using the latter.

Fixed PS1 to not leave detritus when using history (via up and down arrows);
this is caused by using the color variables on the last prompt line; perhaps
if just the bare ANSI escapes are used, then it wouldn't happen; in any
case, no color looks fine, so going with it.